+++ src/common/js/configure.js
@@ -4,9 +4,14 @@
 
 import _ from 'lodash';
 
+import { ReactiveVar } from 'meteor/reactive-var';
+
 EnvSettings._conf = {};
 
+EnvSettings._configured = new ReactiveVar( false );
+
 EnvSettings._defaults = {
+    onReady(){},
     verbosity: EnvSettings.C.Verbose.READY
 };
 
@@ -23,6 +28,7 @@ EnvSettings.configure = function( o ){
     }
     // verbosity management _after_ having set the conf
     _verbose( EnvSettings.C.Verbose.CONFIGURE, 'configure() with', o, 'building', EnvSettings._conf );
+    EnvSettings._configured.set( true );
     // also acts as a getter
     return EnvSettings._conf;
 };